Move obsolete code, clean up
