+++ mdsimulator/neighbor_list.py
@@ -1,6 +1,6 @@
 import numpy as np
 import numpy.testing as npt
-from neighbor_order_pbc import get_n_cells #Punkt!
+from .neighbor_order_pbc import get_n_cells #Punkt!
 
 
 class NeighborList(object):
