+++ mdsimulator/test/test_short_ranged.py
@@ -1,7 +1,7 @@
 import numpy as np
-from short_ranged import potentials
-from neighbor_list import NeighborList
-from neighbor_order_pbc import create_nb_order
+from ..short_ranged import potentials
+from ..neighbor_list import NeighborList
+from ..neighbor_order_pbc import create_nb_order
 import matplotlib.pyplot as plt
 import numpy.testing as npt
